feat(s3): add methods to interface with buckets
